# src/Output/openapi.yaml
openapi: 3.0.0
info:
  title: 'Calendar App'
  description: 'This API is designed to facilitate the management of a comprehensive calendar application, encompassing a broad range of functionalities including user management, event scheduling, course allocation, group coordination, department organization, and timetable structuring. It aims to provide an integrated platform for educational institutions to efficiently organize and access academic and administrative information.'
  version: 1.0.3
servers:
  -
    url: 'http://localhost/web_project/src/'
paths:
...
  /rooms:
    get:
      tags:
        - Rooms
      summary: 'Retrieve all rooms'
      description: 'Returns all rooms available.'
      operationId: 334929f8efcb91e002eec600e3eca265
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: success }
                  data: { type: array, items: { $ref: '#/components/schemas/Room' } }
                type: object
        '404':
          description: 'Rooms not found'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: 'Rooms not found' }
                type: object
    post:
      tags:
        - Rooms
      summary: 'Create a new room'
      description: 'Creates a new room.'
      operationId: 36ee35d0abd5df8d09ab46d1a5503dac
      requestBody:
        description: 'Room data to create'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Room'
      responses:
        '201':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: success }
                  data: { $ref: '#/components/schemas/Room' }
                type: object
        '400':
          description: 'Invalid data'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: 'Invalid data' }
                type: object
  '/rooms/{id}':
    get:
      tags:
        - Rooms
      summary: 'Retrieve a room by ID'
      description: 'Returns a room by its ID.'
      operationId: 83eb275e6131386dd7c2d4469aa36995
      parameters:
        -
          name: id
          in: path
          description: 'ID of the room to retrieve'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: success }
                  data: { $ref: '#/components/schemas/Room' }
                type: object
        '404':
          description: 'Room not found'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: 'Room not found' }
                type: object
    put:
      tags:
        - Rooms
      summary: 'Update a room by ID'
      description: 'Updates a room by its ID.'
      operationId: 3f748f6805163d8c74fc1a729a139d52
      parameters:
        -
          name: id
          in: path
          description: 'ID of the room to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: 'Room data to update'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Room'
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: success }
                  data: { $ref: '#/components/schemas/Room' }
                type: object
        '400':
          description: 'Invalid data'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: 'Invalid data' }
                type: object
        '404':
          description: 'Room not found'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: 'Room not found' }
                type: object
    delete:
      tags:
        - Rooms
      summary: 'Delete a room by ID'
      description: 'Deletes a room by its ID.'
      operationId: a4af362aa530b602b815d92f75897871
      parameters:
        -
          name: id
          in: path
          description: 'ID of the room to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 'Room deleted successfully'
        '404':
          description: 'Room not found'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: 'Room not found' }
                type: object
  /users:
    get:
      tags:
        - Users
      summary: 'Retrieve one user by email or all users'
      description: 'Returns a single user if email is provided or all users if no email is provided.'
      operationId: 0940fc3675fe6a0a70eceb8334255fbf
      parameters:
        -
          name: email
          in: query
          description: 'The email of the user to retrieve'
          required: false
          schema:
            type: string
            format: email
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                oneOf:
                  - { description: 'List of all users', type: array, items: { $ref: '#/components/schemas/User' } }
                  - { $ref: '#/components/schemas/User' }
        '404':
          description: 'User not found'
          content:
            application/json:
              schema:
                properties:
                  code: { type: integer, example: '404' }
                  status: { type: string, example: error }
                  message: { type: string, example: 'User not found' }
                type: object
        '500':
          description: 'Internal Server Error'
    post:
      tags:
        - Users
      summary: 'Create a new user'
      description: 'Creates a new user.'
      operationId: 350e1f8d85209c87247345e2e1491a7c
      requestBody:
        description: 'User data'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: 'User created successfully'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: 'Bad request' }
                type: object
  '/users/{id}':
    get:
      tags:
        - Users
      summary: 'Get a user by ID'
      description: 'Retrieves a user by its ID.'
      operationId: c551a1ac1be37485844fea99fd5c4ec5
      parameters:
        -
          name: id
          in: path
          description: 'ID of the user to retrieve'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: 'User not found'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: 'User not found' }
                type: object
    put:
      tags:
        - Users
      summary: 'Update a user by ID'
      description: 'Updates a user by its ID. Only certain fields can be updated.'
      operationId: 4e73057119289c1852c31b99ead70192
      parameters:
        -
          name: id
          in: path
          description: 'ID of the user to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: 'User object that needs to be updated'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: 'User updated successfully'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: 'Bad request' }
                type: object
        '404':
          description: 'User not found'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: 'User not found' }
                type: object
    delete:
      tags:
        - Users
      summary: 'Delete a user by ID'
      description: 'Deletes a user by its ID.'
      operationId: 2e3cbd202fb5b06e4e385ea633ba7f06
      parameters:
        -
          name: id
          in: path
          description: 'ID of the user to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 'User deleted successfully'
        '404':
          description: 'User not found'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: 'User not found' }
                type: object
components:
  schemas:
...
    Room:
      title: Room
      description: 'Room entity schema'
      required:
        - id
        - name
      properties:
        id:
          description: 'Unique identifier for the room'
          type: integer
          format: int64
        name:
          description: 'Name of the room'
          type: string
      type: object
    User:
      title: User
      description: 'User entity schema'
      required:
        - id
        - email
        - firstName
        - lastName
        - type
      properties:
        id:
          description: 'Unique identifier for the user'
          type: integer
          format: int64
        email:
          description: 'User email address'
          type: string
          format: email
        firstName:
          description: "User's first name"
          type: string
        lastName:
          description: "User's last name"
          type: string
        type:
          description: 'The type of user'
          type: string
          enum:
            - student
            - teacher
            - admin
        password:
          description: 'User password, not generally returned in responses for security reasons'
          type: string
          writeOnly: true
          example: examplePassword123!
      type: object